Inline track_email_delivery back into Notification
With graduated_notification now routing through a Notification record,
Notification is the only consumer of the concern. Inlining removes a
layer of indirection that no longer earns its keep.
